add ResourcesCard component
